Web/Juicy/requirements.txt
Flask==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-DotEnv==0.1.2
Flask-Migrate==2.7.0
PyJWT==1.7.1
Werkzeug==2.0.0
click==8.0.0
itsdangerous==2.0.0
Jinja2==3.0.0
dominate==2.6.0
visitor==0.1.3
Flask-SQLAlchemy==2.5.1
alembic==1.6.2
SQLAlchemy==1.4.15
Mako==1.1.0
python-editor==1.0.4
python-dateutil==2.7.3
MarkupSafe==2.0.0
greenlet==1.1.0
six==1.14.0


